back to basics. Using reusable components.
